+++ lib/src/plugins/less_plugin_clean_css/clean_css_processor.dart
@@ -9,7 +9,7 @@ class CleanCssProcessor extends Processor {
   CleanCssProcessor(cleanCssOptions):super(cleanCssOptions);
 
   String process(String input, Map options) {
-      return input.toUpperCase(); //TODO for test
+      return input;
   }
 }